# src/main/resources/openapi/openapi-geography.yaml
openapi: 3.0.3
info:
  title: Vacapp Geographic Control API
  description: API REST para gestión jerárquica de terrenos ganaderos (Rancho → Sección → Potrero)
  version: 1.0.0
  contact:
    name: Vacapp Team
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo local
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/geography/potreros:
    post:
      tags:
        - Potreros
      summary: Crear nuevo potrero
      operationId: createPotrero
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePotreroRequest'
      responses:
        '201':
          description: Potrero creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PotreroResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    
    get:
      tags:
        - Potreros
      summary: Listar potreros del tenant
      operationId: listPotreros
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Lista de potreros
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PotreroResponse'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/geography/potreros/{id}:
    get:
      tags:
        - Potreros
      summary: Obtener potrero por ID
      operationId: getPotreroById
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: Potrero encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PotreroResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    
    put:
      tags:
        - Potreros
      summary: Actualizar potrero
      operationId: updatePotrero
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePotreroRequest'
      responses:
        '200':
          description: Potrero actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PotreroResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    
    delete:
      tags:
        - Potreros
      summary: Archivar potrero
      operationId: archivePotrero
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '204':
          description: Potrero archivado exitosamente
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPathParam:
      name: id
      in: path
      required: true
      description: UUID del recurso
      schema:
        type: string
        format: uuid
  schemas:
...
    CreatePotreroRequest:
      type: object
      required:
        - nombre
        - superficie
        - ranchoId
      properties:
        nombre:
          type: string
          minLength: 2
          maxLength: 100
          example: "Potrero A1"
        superficie:
          type: number
          format: decimal
          minimum: 0.01
          maximum: 999999999
          example: 5000.00
        ranchoId:
          type: string
          format: uuid
        seccionId:
          type: string
          format: uuid
          nullable: true
          description: Opcional - vincular a una sección
        descripcion:
          type: string
          maxLength: 500
    UpdatePotreroRequest:
      type: object
      required:
        - nombre
        - superficie
      properties:
        nombre:
          type: string
          minLength: 2
          maxLength: 100
        superficie:
          type: number
          format: decimal
          minimum: 0.01
          maximum: 999999999
        descripcion:
          type: string
          maxLength: 500
    
    # ===== RESPONSE DTOs =====
...
    PotreroResponse:
      type: object
      properties:
        potreroId:
          type: string
          format: uuid
        nombre:
          type: string
        superficie:
          type: number
          format: decimal
        ranchoId:
          type: string
          format: uuid
        seccionId:
          type: string
          format: uuid
          nullable: true
        cattleCount:
          type: integer
          minimum: 0
          description: Cantidad de ganado asignado
        descripcion:
          type: string
        status:
          type: string
          enum: [ACTIVE, ARCHIVED]
        tenantId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          minimum: 0
          description: Página actual (0-based)
        size:
          type: integer
          minimum: 1
          description: Tamaño de página
        total:
          type: integer
          minimum: 0
          description: Total de elementos
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  responses:
    BadRequest:
      description: Petición inválida (validación fallida)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Autenticación requerida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflicto (ej. nombre duplicado)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'